backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
redis
celery
python-dotenv
email-validator
requests
reportlab
prometheus-fastapi-instrumentator
python-logstash-async
certifi
